requirements.txt
aiohttp>=3.9.0
aiohttp-socks>=0.8.4
dnspython>=2.4.0
python-whois>=0.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rich>=13.7.0
pyperclip>=1.8.2
